cleanup, refactor, and add doc comments
