+++ src/vs/workbench/contrib/chat/common/promptSyntax/languageFeatures/providers/index.ts
@@ -6,10 +6,10 @@
 import { PromptLinkProvider } from './promptLinkProvider.js';
 import { isWindows } from '../../../../../../../base/common/platform.js';
 import { PromptPathAutocompletion } from './promptPathAutocompletion.js';
-import { PromptDecoratorsInstanceManager } from './textModelPromptDecorator.js';
 import { Registry } from '../../../../../../../platform/registry/common/platform.js';
 import { LifecyclePhase } from '../../../../../../services/lifecycle/common/lifecycle.js';
 import { PromptLinkDiagnosticsInstanceManager } from './promptLinkDiagnosticsProvider.js';
+import { PromptDecorationsProviderInstanceManager } from './decorationsProvider/promptDecorationsProvider.js';
 import { BrandedService } from '../../../../../../../platform/instantiation/common/instantiation.js';
 import { IWorkbenchContributionsRegistry, Extensions, IWorkbenchContribution } from '../../../../../../common/contributions.js';
 
@@ -18,8 +18,8 @@ import { IWorkbenchContributionsRegistry, Extensions, IWorkbenchContribution } f
  */
 export const registerReusablePromptLanguageFeatures = () => {
 	registerContribution(PromptLinkProvider);
-	registerContribution(PromptDecoratorsInstanceManager);
 	registerContribution(PromptLinkDiagnosticsInstanceManager);
+	registerContribution(PromptDecorationsProviderInstanceManager);
 
 	/**
 	 * We restrict this provider to `Unix` machines for now because of
@@ -42,5 +42,4 @@ const registerContribution = <TServices extends BrandedService[]>(
 ) => {
 	Registry.as<IWorkbenchContributionsRegistry>(Extensions.Workbench)
 		.registerWorkbenchContribution(contribution, LifecyclePhase.Eventually);
-
 };
